% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesDeming-package.R
\docType{package}
\name{BayesDeming-package}
\alias{BayesDeming-package}
\alias{BayesDeming}
\title{The 'BayesDeming' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2022). RStan: the R interface to Stan. R package version 2.21.5. https://mc-stan.org
}
