% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_data}
\alias{plot_data}
\title{Plot Deming data}
\usage{
plot_data(
  df,
  formula = y ~ x,
  rug = TRUE,
  elipses = TRUE,
  a = NULL,
  b = NULL,
  fixed = TRUE
)
}
\description{
Plot Deming data
}
\examples{
df <- sim_data(
  Jx = c(5, 5, 5, 5, 5, 7, 4, 4),
  Jy = c(6, 6, 6, 6, 6, 6, 6, 6),
  sdx = sqrt(((1/50)/2)^.5), lambda = 2
)

plot_data(df)
}
