% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-data.R
\name{sim_data}
\alias{sim_data}
\title{Simulate Deming data}
\usage{
sim_data(
  Jx,
  Jy,
  theta = NULL,
  sdx,
  sdy = NULL,
  lambda = NULL,
  alpha = 0,
  beta = 1
)
}
\description{
Simulate Deming data
}
\examples{
df <- sim_data(
  Jx = c(5, 5, 5, 5, 5, 7, 4, 4),
  Jy = c(6, 6, 6, 6, 6, 6, 6, 6),
  sdx = sqrt(((1/50)/2)^.5), lambda = 2
)
}
