% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deming.R
\name{deming}
\alias{deming}
\title{Bayesian linear regression with Stan}
\usage{
deming(df, priors, formula = y ~ x, z = NULL, power = NULL, ...)
}
\arguments{
\item{df}{Nested tibble, output from \code{sim_data}}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Bayesian linear regression with Stan
}
