% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-data.R
\name{sim_data}
\alias{sim_data}
\title{Simulate Deming data}
\usage{
sim_data(
  Jx,
  Jy,
  theta = NULL,
  sdx,
  sdy = NULL,
  lambda = NULL,
  alpha = 0,
  beta = 1
)
}
\arguments{
\item{Jx, Jy}{numeric vectors of equal lengths, the number of repeated measures per group.}

\item{theta}{numeric vector the same length as \code{Jx} and \code{Jy}, the X values group means}

\item{sdx, sdy}{the standard deviation of the X and Y values across all groups}

\item{lambda}{the ratio of \code{sdy}/\code{sdx}. Either \code{sdy} or \code{lambda} must be specified, but not both.}

\item{alpha, beta}{the values of the intercept and slope relating X and Y group means.}
}
\description{
Simulate Deming data
}
\examples{
df <- sim_data(
  Jx = c(5, 5, 5, 5, 5, 7, 4, 4),
  Jy = c(6, 6, 6, 6, 6, 6, 6, 6),
  sdx = sqrt(((1/50)/2)^.5), lambda = 2
)
}
